requirements.txt
pandas==2.1.4
numpy==1.26.2
mysql-connector-python==8.2.0
reportlab==4.0.7
openpyxl==3.1.2
